+++ src/main/java/com/ikkiking/api/response/PostResponse/Comment.java
@@ -3,23 +3,23 @@ package com.ikkiking.api.response.PostResponse;
 import java.util.Date;
 
 public class Comment {
-    private Integer id;
+    private Long id;
     private long timestamp;
     private String text;
     private CommentUser user;
 
-    public Comment(Integer id, long timestamp, String text, CommentUser user) {
+    public Comment(Long id, long timestamp, String text, CommentUser user) {
         this.id = id;
         this.timestamp = timestamp;
         this.text = text;
         this.user = user;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
