+++ src/main/java/com/ikkiking/api/response/PostResponse/User.java
@@ -2,19 +2,19 @@ package com.ikkiking.api.response.PostResponse;
 
 public class User {
 
-    private Integer id;
+    private Long id;
     private String name;
 
-    public User(Integer id, String name) {
+    public User(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
