+++ src/main/java/com/ikkiking/api/response/StatisticResponse/AllStatisticResponse.java
@@ -1,7 +1,7 @@
 package com.ikkiking.api.response.StatisticResponse;
 
 public class AllStatisticResponse extends StatisticResponse{
-    public AllStatisticResponse(int postsCount, int likesCount, int dislikesCount, int viewsCount, long firstPublication) {
+    public AllStatisticResponse(long postsCount, long likesCount, long dislikesCount, long viewsCount, long firstPublication) {
         super(postsCount, likesCount, dislikesCount, viewsCount, firstPublication);
     }
 }
